requirements.txt
joblib
torch>=1.8.0
matplotlib
pandas
numpy
tensorboardX
scikit-learn
scipy
kaleido
psutil
tqdm
biopython